Structural fixes & updated error logs for HPP
